+++ changelog.md
@@ -19,6 +19,7 @@
   - Don’t specify the `CI` environment variable as it’s set by default.
 - Use a new [`@arr/flatten`](https://npm.im/@arr/flatten) dev dependency to flatten arrays in tests.
 - Stop using [`hard-rejection`](https://npm.im/hard-rejection) to detect unhandled `Promise` rejections in tests, as Node.js v15+ does this natively.
+- Added a test for `FirstRenderDateContext` used as a React context.
 
 ## 12.0.1